back/docprocessor/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

python-dotenv==1.0.1
requests==2.32.3

azure-functions==1.22.1
azure-functions-durable==1.2.10

langchain==0.3.7
langchain-community==0.3.7

mistralai==1.5.1

azure-storage-blob==12.19.1
azure-cosmos==4.9.0

pytest==8.3.5


